feat: zod validation api equipment

+++ app/api/equipment/[id]/route.ts
@@ -1,6 +1,9 @@
 import Connect from '@/lib/mongodb/connect';
 import { ObjectId } from 'mongodb';
 import { NextResponse } from 'next/server';
+import { z } from 'zod';
+
+const idSchema = z.string();
 
 export interface IEquipment {
   params: { id: string };
@@ -8,6 +11,17 @@ export interface IEquipment {
 
 export async function GET(request: Request, { params }: IEquipment) {
   const { id } = params;
+
+  // validation
+  const validation = idSchema.safeParse(id);
+  if (!validation.success) {
+    const { errors } = validation.error;
+
+    return NextResponse.json(
+      { message: 'Invalid request', errors },
+      { status: 400 }
+    );
+  }
 
   const collection = await Connect('equipment');
   const response = await collection.findOne({
@@ -19,6 +33,17 @@ export async function GET(request: Request, { params }: IEquipment) {
 
 export async function PUT(request: Request, { params }: IEquipment) {
   const { id } = params;
+
+  // validation
+  const validation = idSchema.safeParse(id);
+  if (!validation.success) {
+    const { errors } = validation.error;
+
+    return NextResponse.json(
+      { message: 'Invalid request', errors },
+      { status: 400 }
+    );
+  }
 
   const { name, category, location } = await request.json();
   const collection = await Connect('equipment');
@@ -39,6 +64,17 @@ export async function PUT(request: Request, { params }: IEquipment) {
 export async function DELETE(request: Request, { params }: IEquipment) {
   const { id } = params;
 
+  // validation
+  const validation = idSchema.safeParse(id);
+  if (!validation.success) {
+    const { errors } = validation.error;
+
+    return NextResponse.json(
+      { message: 'Invalid request', errors },
+      { status: 400 }
+    );
+  }
+
   const collection = await Connect('equipment');
   const response = await collection.deleteOne({ _id: new ObjectId(id) });
 
@@ -48,6 +84,17 @@ export async function DELETE(request: Request, { params }: IEquipment) {
 // DELETING WITH PATCH (DELETE WORKS WITH POSTMAN BUT NOT IN NEXTJS)
 export async function PATCH(request: Request, { params }: IEquipment) {
   const { id } = params;
+
+  // validation
+  const validation = idSchema.safeParse(id);
+  if (!validation.success) {
+    const { errors } = validation.error;
+
+    return NextResponse.json(
+      { message: 'Invalid request', errors },
+      { status: 400 }
+    );
+  }
 
   const collection = await Connect('equipment');
   const response = await collection.deleteOne({ _id: new ObjectId(id) });
